+++ cmd/helper_rest.go
@@ -28,6 +28,8 @@ import (
 	"net/http"
 	"time"
 
+	b64 "encoding/base64"
+
 	"github.com/hashicorp/go-slug"
 	"github.com/hashicorp/go-tfe"
 )
@@ -180,4 +182,118 @@ func ListGPGKeys(token string, tfeHostname string, tfeOrganization string) (*GPG
 	}
 
 	return keys, nil
+}
+
+type TFEBinaries struct {
+	Releases []TFERelease
+}
+
+type TFERelease struct {
+	ReleaseSequence      int       `json:"release_sequence"`
+	PatchReleaseSequence int       `json:"patch_release_sequence"`
+	ReleaseDate          time.Time `json:"release_date"`
+	Required             bool      `json:"required"`
+	Label                string    `json:"label"`
+	ReleaseNotes         string    `json:"release_notes"`
+	DownloadLink         string    `json:"download_link"`
+	BuildStatus          string    `json:"build_status"`
+	LastUpdateTime       time.Time `json:"last_update_time"`
+	Checksum             string    `json:"checksum"`
+	ImagelessChecksum    string    `json:"imageless_checksum"`
+}
+
+func ListTFEBinaries(password string, licenseId string) (*TFEBinaries, error) {
+	passwordB64 := b64.URLEncoding.EncodeToString([]byte(password))
+	// create url "https://api.replicated.com/market/v1/airgap/releases?license_id=${LICENSE_ID}"
+	url := fmt.Sprintf(
+		"https://api.replicated.com/market/v1/airgap/releases?license_id=%s",
+		licenseId,
+	)
+	// create http Client to make calls
+	client := &http.Client{}
+
+	// create request
+	req, err := http.NewRequest("GET", url, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	// add headers
+	req.Header.Set("Authorization", "Basic "+passwordB64)
+	req.Header.Set("Accept", "application/json")
+
+	// make request
+	resp, err := client.Do(req)
+	if err != nil {
+		return nil, err
+	}
+
+	// wait for complete
+	defer resp.Body.Close()
+
+	// read all bytes, convert to object
+	bodyBytes, _ := ioutil.ReadAll(resp.Body)
+
+	// bodyString := string(bodyBytes)
+	// fmt.Println("API Response as String:\n" + bodyString)
+
+	// Convert response body to Todo struct
+	var tfeBinaries *TFEBinaries
+	err = json.Unmarshal(bodyBytes, &tfeBinaries)
+	if err != nil {
+		return nil, err
+	}
+
+	return tfeBinaries, nil
+}
+
+type TfeReleaseUrl struct {
+	URL          string `json:"url"`
+	ImagelessURL string `json:"imageless_url"`
+}
+
+func GetTFEBinary(password string, licenseId string, releaseSequence string) (*TfeReleaseUrl, error) {
+	passwordB64 := b64.URLEncoding.EncodeToString([]byte(password))
+	// create url "https://api.replicated.com/market/v1/airgap/images/url?license_id=${LICENSE_ID}&sequence=${release_sequence}"
+	url := fmt.Sprintf(
+		"https://api.replicated.com/market/v1/airgap/images/url?license_id=%s&sequence=%s",
+		licenseId,
+		releaseSequence,
+	)
+	// create http Client to make calls
+	client := &http.Client{}
+
+	// create request
+	req, err := http.NewRequest("GET", url, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	// add headers
+	req.Header.Set("Authorization", "Basic "+passwordB64)
+	req.Header.Set("Accept", "application/json")
+
+	// make request
+	resp, err := client.Do(req)
+	if err != nil {
+		return nil, err
+	}
+
+	// wait for complete
+	defer resp.Body.Close()
+
+	// read all bytes, convert to object
+	bodyBytes, _ := ioutil.ReadAll(resp.Body)
+
+	// bodyString := string(bodyBytes)
+	// fmt.Println("API Response as String:\n" + bodyString)
+
+	// Convert response body to Todo struct
+	var tfeUrl *TfeReleaseUrl
+	err = json.Unmarshal(bodyBytes, &tfeUrl)
+	if err != nil {
+		return nil, err
+	}
+
+	return tfeUrl, nil
 }